Tile details are shown while hovering the mouse cursor instead of having to click.
Fixes #458.

Before you had to click on the tile in an overmap to get the tile details, now you can simply hover and it'll automatically show the tile details to the player.

I didn't remove the input signal, since it can now be used for something else.

Note: this is only applied to in-game map (I tested it there, not in content editor).